+++ emp_profile/views/views.xml
@@ -173,39 +173,7 @@
     <menuitem name="Year Of Company" id="wizard_company_menu_2_year" parent="emp_profile.menu_2"
               action="wizard_company_action_window"/>
 
-    <!-- Department -->
-
-    <record id="department_types_form" model="ir.ui.view">
-      <field name="name">department.types.form</field>
-      <field name="model">department.types</field>
-      <field name="arch" type="xml">
-        <form>
-          <group>
-            <field name="name"  />
-          </group>
-        </form>
-      </field>
-    </record>
-
-    <record id="department_types_list" model="ir.ui.view">
-      <field name="name">department.types.list</field>
-      <field name="model">department.types</field>
-      <field name="arch" type="xml">
-        <tree editable='bottom'>
-          <field name="sequence" widget="handle"/>
-          <field name="id" />
-          <field name="name" required='1' />
-        </tree>
-      </field>
-    </record>
-
-    <record model="ir.actions.act_window" id="department_types_action_window">
-      <field name="name">employee profile Department</field>
-      <field name="res_model">department.types</field>
-      <field name="view_mode">tree,form</field>
-    </record>
-
-    <menuitem id="menu_employee_department" name="Department" parent="emp_profile.menu_1" action="emp_profile.department_types_action_window" />
+    
 
   </data>
 </odoo>
